requirements.txt
fastapi
uvicorn[standard]
celery[redis]
python-dotenv
google-generativeai
pinecone
redis
tenacity
requests